# magazine_api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.UserProfile:
    properties:
      contact_number:
        type: string
      created_on:
        type: string
      deleted_on:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      picture:
        type: string
      updated_on:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:5005
info:
  contact:
    name: 21557870 - Suman Chhetri
  description: This is an API documentation of Magazine API
  title: Magazine API Documentation
  version: v1
paths:
...
  /user_profile:
    post:
      consumes:
      - application/json
      description: It creates UserProfile
      parameters:
      - description: Add UserProfile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.UserProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.UserProfile'
              type: object
      summary: Create UserProfile
      tags:
      - UserProfile
  /user_profile/{id}:
    delete:
      description: Delete by user_profile ID
      parameters:
      - description: Unique ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  type: string
              type: object
      summary: Soft Delete an user_profile
      tags:
      - UserProfile
    patch:
      consumes:
      - application/json
      description: Updates UserProfile of user
      parameters:
      - description: Update user_profile
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.UserProfile'
              type: object
      summary: Update UserProfile
      tags:
      - UserProfile
  /user_profile/id/{id}:
    get:
      description: Gets One user profile by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
      summary: Gets One User Profile by ID
      tags:
      - UserProfile
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"